[megatron] fix: critic and reward model load tokenizer from config (#301)
Currently, the worker will fail if the critic or reward model path
doesn't contain a tokenizer. This PR tries to fix this by loading
tokenizer from the config for the previously mentioned case.

- For the critic model, we fall back to load from
`critic.model.tokenizer_path`.
- For the reward model, we first fall back to load from
`reward_model.model.rm_tokenizer`, throw an error if that is not set.

---------

Signed-off-by: Hollow Man <hollowman@opensuse.org>
Co-authored-by: ETOgaosion <gaoziyuan19@mails.ucas.ac.cn>